Added basic support for SOME of the floating insts

+++ src/main/kotlin/venus/simulator/diffs/FRegisterDiff.kt
@@ -3,6 +3,6 @@ package venus.simulator.diffs
 import venus.simulator.Diff
 import venus.simulator.SimulatorState
 
-class FRegisterDiff(val id: Int, val v: Int) : Diff {
+class FRegisterDiff(val id: Int, val v: Float) : Diff {
     override operator fun invoke(state: SimulatorState) = state.setFReg(id, v)
 }
